test: move sweep mocks to TestSupport, rename tests, tag concurrency
Mocks of public protocols (AssociatedTokenAccountFetching, StatelessSwapping)
and the TokenAccountInfo fixture move out of the test file into TestSupport
following the Mock<Protocol>.swift convention. Existing tests rename to
the methodName_scenario_expectedResult convention. Actor-isolation tests
gain .tags(.concurrency) so they can be filtered under TSan in CI.

The Counter and Expectation helpers stay inline — only one test file uses
them, so a TestSupport move would be premature.
