+++ FlipcashTests/UsdcSweepOperationTests.swift
@@ -14,10 +14,13 @@ struct UsdcSweepOperationTests {
 
     // MARK: - Single in-flight
 
-    @Test("Re-entrant start() calls don't double-fire — the second is skipped while the first is still awaiting the fetch")
-    func reentrantStartSkipsWhenInFlight() async throws {
-        let fetcher = MockAccountFetcher()
-        let swapper = MockSwapper()
+    @Test(
+        "Re-entrant start() calls don't double-fire — the second is skipped while the first is still awaiting the fetch",
+        .tags(.concurrency)
+    )
+    func start_whileInFlight_skipsSecondCall() async throws {
+        let fetcher = MockAssociatedTokenAccountFetching()
+        let swapper = MockStatelessSwapping()
         let completion = Counter()
 
         let op = UsdcSweepOperation(
@@ -60,9 +63,9 @@ struct UsdcSweepOperationTests {
         "Empty USDC ATA short-circuits before the swap",
         arguments: [Optional<UInt64>.none, .some(0)]
     )
-    func emptyAtaSkipsSwap(quarks: UInt64?) async throws {
-        let fetcher = MockAccountFetcher()
-        let swapper = MockSwapper()
+    func start_whenAtaEmpty_skipsSwap(quarks: UInt64?) async throws {
+        let fetcher = MockAssociatedTokenAccountFetching()
+        let swapper = MockStatelessSwapping()
         let completion = Counter()
 
         let op = UsdcSweepOperation(
@@ -87,8 +90,8 @@ struct UsdcSweepOperationTests {
 
     @Test("Positive USDC balance runs the swap and fires onSweepCompleted")
     func start_withPositiveBalance_invokesSwapAndCompletion() async throws {
-        let fetcher = MockAccountFetcher()
-        let swapper = MockSwapper()
+        let fetcher = MockAssociatedTokenAccountFetching()
+        let swapper = MockStatelessSwapping()
         let completion = Counter()
 
         let op = UsdcSweepOperation(
@@ -110,10 +113,13 @@ struct UsdcSweepOperationTests {
 
     // MARK: - Re-entry after completion
 
-    @Test("Second start() after a successful sweep runs another sweep")
+    @Test(
+        "Second start() after a successful sweep runs another sweep",
+        .tags(.concurrency)
+    )
     func start_afterPreviousCompletion_allowsSecondSweep() async throws {
-        let fetcher = MockAccountFetcher()
-        let swapper = MockSwapper()
+        let fetcher = MockAssociatedTokenAccountFetching()
+        let swapper = MockStatelessSwapping()
         let completion = Counter()
 
         let op = UsdcSweepOperation(
@@ -143,10 +149,13 @@ struct UsdcSweepOperationTests {
 
     // MARK: - Error path
 
-    @Test("Swap failure skips the completion callback and releases the in-flight guard")
+    @Test(
+        "Swap failure skips the completion callback and releases the in-flight guard",
+        .tags(.concurrency)
+    )
     func start_whenSwapThrows_skipsCompletionAndReleasesGuard() async throws {
-        let fetcher = MockAccountFetcher()
-        let swapper = MockSwapper()
+        let fetcher = MockAssociatedTokenAccountFetching()
+        let swapper = MockStatelessSwapping()
         let completion = Counter()
 
         let op = UsdcSweepOperation(
@@ -172,70 +181,7 @@ struct UsdcSweepOperationTests {
     }
 }
 
-// MARK: - Mocks
-
-private actor MockAccountFetcher: AssociatedTokenAccountFetching {
-
-    private(set) var callCount = 0
-
-    private enum Mode {
-        case immediate(@Sendable () throws -> AccountInfo?)
-        case blocking(@Sendable () -> Void)
-    }
-    private var mode: Mode?
-    private var pendingContinuation: CheckedContinuation<Result<AccountInfo?, Error>, Never>?
-
-    func setImmediateHandler(_ handler: @escaping @Sendable () throws -> AccountInfo?) {
-        mode = .immediate(handler)
-    }
-
-    func setBlockingHandler(_ onEntered: @escaping @Sendable () -> Void) {
-        mode = .blocking(onEntered)
-    }
-
-    func resumeWith(_ result: Result<AccountInfo?, Error>) {
-        pendingContinuation?.resume(returning: result)
-        pendingContinuation = nil
-    }
-
-    func fetchAssociatedTokenAccount(owner: KeyPair, mint: PublicKey) async throws -> AccountInfo? {
-        callCount += 1
-        switch mode {
-        case .immediate(let handler):
-            return try handler()
-        case .blocking(let onEntered):
-            onEntered()
-            let result: Result<AccountInfo?, Error> = await withCheckedContinuation { c in
-                pendingContinuation = c
-            }
-            return try result.get()
-        case nil:
-            return nil
-        }
-    }
-}
-
-private actor MockSwapper: StatelessSwapping {
-
-    private(set) var callCount = 0
-    private var nextResult: Result<StatelessSwapResult, Error> = .success(
-        .finalized(signature: try! Signature(Data(repeating: 0, count: 64)))
-    )
-
-    func setNextResult(_ result: Result<StatelessSwapResult, Error>) {
-        nextResult = result
-    }
-
-    func statelessSwap(
-        fromMint: MintMetadata,
-        toMint: MintMetadata,
-        amount: TokenAmount,
-        owner: KeyPair
-    ) async throws -> StatelessSwapResult {
-        callCount += 1
-        return try nextResult.get()
-    }
-}
+// MARK: - Local helpers
 
 private actor Counter {
     private(set) var value = 0
@@ -274,21 +220,5 @@ private final class Expectation: @unchecked Sendable {
                 lock.unlock()
             }
         }
-    }
-}
-
-// MARK: - Fixtures
-
-private extension Ocp_Account_V1_TokenAccountInfo {
-    static func usdcAtaInfo(quarks: UInt64) -> Ocp_Account_V1_TokenAccountInfo {
-        let id = PublicKey.mock.solanaAccountID
-        var proto = Ocp_Account_V1_TokenAccountInfo()
-        proto.address = id
-        proto.mint = id
-        proto.owner = id
-        proto.authority = id
-        proto.balance = quarks
-        proto.accountType = .associatedTokenAccount
-        return proto
     }
 }
